# contracts/generated/openapi/contracts.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiskItem ModelOps API",
    "version": "0.1.0",
    "description": "信贷风控模型智能监测与自主迭代系统 API。由 scripts/generate_openapi.py 从 FastAPI app 自动生成，不手工维护。"
  },
  "paths": {
...
    "/api/iteration/data-eligibility": {
      "post": {
        "tags": [
          "iteration"
        ],
        "summary": "Evaluate Data Eligibility",
        "operationId": "evaluate_data_eligibility_api_iteration_data_eligibility_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataEligibilityInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iteration/decisions": {
      "post": {
        "tags": [
          "iteration"
        ],
        "summary": "Create Decision",
        "operationId": "create_decision_api_iteration_decisions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "iteration"
        ],
        "summary": "List Decisions",
        "operationId": "list_decisions_api_iteration_decisions_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iteration/decisions/{proposal_id}": {
      "get": {
        "tags": [
          "iteration"
        ],
        "summary": "Get Decision",
        "operationId": "get_decision_api_iteration_decisions__proposal_id__get",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iteration/proposals/{proposal_id}": {
      "get": {
        "tags": [
          "iteration"
        ],
        "summary": "Get Proposal Contract",
        "operationId": "get_proposal_contract_api_iteration_proposals__proposal_id__get",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iteration/proposals/{proposal_id}/risk": {
      "get": {
        "tags": [
          "iteration"
        ],
        "summary": "Get Proposal Risk",
        "operationId": "get_proposal_risk_api_iteration_proposals__proposal_id__risk_get",
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iteration/plans/{training_plan_id}": {
      "get": {
        "tags": [
          "iteration"
        ],
        "summary": "Get Training Plan",
        "operationId": "get_training_plan_api_iteration_plans__training_plan_id__get",
        "parameters": [
          {
            "name": "training_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Training Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iteration/runs/{iteration_run_id}": {
      "get": {
        "tags": [
          "iteration"
        ],
        "summary": "Get Iteration Run",
        "operationId": "get_iteration_run_api_iteration_runs__iteration_run_id__get",
        "parameters": [
          {
            "name": "iteration_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iteration Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/iteration/runs/{iteration_run_id}/rounds": {
      "get": {
        "tags": [
          "iteration"
        ],
        "summary": "Get Iteration Rounds",
        "operationId": "get_iteration_rounds_api_iteration_runs__iteration_run_id__rounds_get",
        "parameters": [
          {
            "name": "iteration_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Iteration Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DataEligibilityInput": {
        "properties": {
          "window_id": {
            "type": "string",
            "title": "Window Id"
          },
          "data_track": {
            "$ref": "#/components/schemas/DataTrack",
            "default": "NATURAL"
          },
          "data_snapshot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Snapshot Id"
          },
          "data_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Checksum"
          },
          "label_column": {
            "type": "string",
            "title": "Label Column",
            "default": "is_bad"
          },
          "label_missing_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Label Missing Rate"
          },
          "label_mature": {
            "type": "boolean",
            "title": "Label Mature",
            "default": true
          },
          "label_imputation_requested": {
            "type": "boolean",
            "title": "Label Imputation Requested",
            "default": false
          },
          "feature_missing_stats": {
            "items": {
              "$ref": "#/components/schemas/FeatureMissingStat"
            },
            "type": "array",
            "title": "Feature Missing Stats"
          },
          "requested_for_supervised_training": {
            "type": "boolean",
            "title": "Requested For Supervised Training",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "window_id",
          "label_missing_rate"
        ],
        "title": "DataEligibilityInput"
      },
      "DataTrack": {
        "type": "string",
        "enum": [
          "NATURAL",
          "SCENARIO"
        ],
        "title": "DataTrack"
      },
...
    }
  }
}